--- src/components/List.js
@@ -1,13 +1,0 @@
-import React, {useState} from 'react';
-import axios from 'axios';
-
-
-const List = () => {
-    return (
-        <>
-            <h3>THis is the List page</h3>
-        </>
-    )
-}
-
-export default List;
