+++ src/components/auth/Registration.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import { Form, Field, withFormik, Formik } from "formik";
+import { Form, Field, withFormik, Formik } from "formik/dist/index";
 import * as Yup from "yup";
 import { connect } from "react-redux";
-import { registerUser } from "../actions/authActions";
+import { registerUser } from "../../actions/authActions";
 
 import "./Login.css";
