+++ src/webpack.config.js
@@ -1,0 +1,4 @@
+{
+    test: /\.svg$/,
+        loader: 'svg-inline-loader'
+}
